@@ -7,7 +7,7 @@
 """Provides a configuration interface for the test generator."""
 import dataclasses
 import enum
-from typing import List, Optional
+import time
 
 from pynguin.utils.statistics.runtimevariable import RuntimeVariable
 
@@ -175,14 +175,14 @@ class StatisticsOutputConfiguration:
     timeline_interpolation: bool = True
     """Interpolate timeline values"""
 
-    coverage_metrics: List[CoverageMetric] = dataclasses.field(
+    coverage_metrics: list[CoverageMetric] = dataclasses.field(
         default_factory=lambda: [
             CoverageMetric.BRANCH,
         ]
     )
     """List of coverage metrics that are optimised during the search"""
 
-    output_variables: List[RuntimeVariable] = dataclasses.field(
+    output_variables: list[RuntimeVariable] = dataclasses.field(
         default_factory=lambda: [
             RuntimeVariable.TargetModule,
             RuntimeVariable.Coverage,
@@ -247,7 +247,7 @@ class TestCaseOutputConfiguration:
 class SeedingConfiguration:
     """Configuration related to seeding."""
 
-    seed: Optional[int] = None
+    seed: int = time.time_ns()
     """A predefined seed value for the random number generator that is used."""
 
     constant_seeding: bool = True
@@ -364,7 +364,7 @@ class TypeInferenceConfiguration:
     """The maximum level of recursion when calculating the dependencies in the test
     cluster."""
 
-    stub_dir: Optional[str] = None
+    stub_dir: str = ""
     """Path to the pyi-stub files for the StubInferenceStrategy"""
 
 
