@@ -5,7 +5,7 @@
 #  SPDX-License-Identifier: LGPL-3.0-or-later
 #
 import os
-from typing import Dict, List, Set, Type, cast
+from typing import cast
 
 import pytest
 from ordered_set import OrderedSet
@@ -24,7 +24,7 @@ from pynguin.utils.generic.genericaccessibleobject import (
 )
 
 
-def convert_to_str_count_dict(dic: Dict[Type, OrderedSet]) -> Dict[str, int]:
+def convert_to_str_count_dict(dic: dict[type, OrderedSet]) -> dict[str, int]:
     return {k.__name__: len(v) for k, v in dic.items()}
 
 
@@ -84,12 +84,14 @@ def test_simple_dependencies_only_own_classes():
     assert len(cluster.accessible_objects_under_test) == 1
 
 
-def test_resolve_only_union():
+def test_resolve_dependenices():
     cluster = TestClusterGenerator(
         "tests.fixtures.cluster.typing_parameters"
     ).generate_cluster()
+    # 3 functions
     assert len(cluster.accessible_objects_under_test) == 3
-    assert len(cluster.generators) == 1
+    # 3 non primitive types in args
+    assert len(cluster.generators) == 3
 
 
 def test_resolve_optional():
@@ -152,7 +154,7 @@ def test_overridden_inherited_methods():
 
 def _extract_method_names(
     accessible_objects: OrderedSet[GenericAccessibleObject],
-) -> Set[str]:
+) -> set[str]:
     return {
         f"{elem.owner.__name__}.{elem.callable.__name__}"
         if isinstance(elem, GenericMethod)
@@ -176,7 +178,7 @@ def test_conditional_import_forward_ref():
 def test_enums():
     cluster = TestClusterGenerator("tests.fixtures.cluster.enums").generate_cluster()
     accessible_objects = cast(
-        List[GenericEnum], list(cluster.accessible_objects_under_test)
+        list[GenericEnum], list(cluster.accessible_objects_under_test)
     )
     assert {enum.owner.__name__: set(enum.names) for enum in accessible_objects} == {
         "Color": {"RED", "BLUE", "GREEN"},
