@@ -25,7 +25,7 @@ def test_merge_full():
     trace0.true_distances[1] = 3
     trace0.false_distances[0] = 0
     trace0.false_distances[1] = 1
-    trace0.covered_lines = {0}
+    trace0.covered_line_ids = {0}
 
     trace1 = ExecutionTrace()
     trace1.executed_code_objects.add(1)
@@ -36,7 +36,7 @@ def test_merge_full():
     trace1.true_distances[2] = 3
     trace1.false_distances[1] = 234
     trace1.false_distances[2] = 0
-    trace1.covered_lines = {1}
+    trace1.covered_line_ids = {1}
 
     result = ExecutionTrace()
     result.executed_code_objects.add(0)
@@ -51,7 +51,7 @@ def test_merge_full():
     result.false_distances[0] = 0
     result.false_distances[1] = 1
     result.false_distances[2] = 0
-    result.covered_lines = {0, 1}
+    result.covered_line_ids = {0, 1}
 
     trace0.merge(trace1)
     assert trace0 == result
